Reorganize config classes & add enableAdminEndpoints option
