+++ bot/src/main/kotlin/org/epilink/bot/config/LinkConfig.kt
@@ -12,9 +12,8 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
 import com.fasterxml.jackson.module.kotlin.KotlinModule
 import org.epilink.bot.CliArgs
-import org.epilink.bot.rulebook.Rulebook
-import org.epilink.bot.discord.DiscordEmbed
 import org.epilink.bot.discord.StandardRoles
+import org.epilink.bot.rulebook.Rulebook
 import java.nio.file.Files
 import java.nio.file.Path
 
@@ -79,312 +78,12 @@ data class LinkConfiguration(
     val admins: List<String> = listOf()
 )
 
-/**
- * Used in [LinkWebServerConfiguration.proxyType] to set the kind of proxy expected
- */
-enum class ProxyType {
-    /**
-     * No proxy used: do not enable header support
-     */
-    None,
-
-    /**
-     * Proxy used with the de-facto standard X-Forwarded-* headers
-     */
-    XForwarded,
-
-    /**
-     * Proxy used with the standard Forwarded header
-     */
-    Forwarded
-}
-
-/**
- * HTTP/Web server options
- *
- * @see LinkConfiguration.server
- */
-data class LinkWebServerConfiguration(
-    /**
-     * The port to be used by the server
-     */
-    val port: Int,
-    /**
-     * Determine which (possibly de-facto) standard to follow for proxy headers support.
-     */
-    val proxyType: ProxyType,
-    /**
-     * The URL to the front-end, with a trailing slash, or null if the front-end is bundled
-     */
-    val frontendUrl: String?,
-    /**
-     * A list of footers that should be displayed in the front-end
-     */
-    val footers: List<LinkFooterUrl> = listOf(),
-    /**
-     * Logo resource asset that is passed to the front-end
-     */
-    val logo: ResourceAssetConfig? = ResourceAssetConfig(),
-    /**
-     * Background resource asset that is passed to the front-end
-     */
-    val background: ResourceAssetConfig? = ResourceAssetConfig(),
-    /**
-     * Contact information for instance maintainers
-     *
-     * @since 0.2.0
-     */
-    val contacts: List<LinkContactInformation> = listOf()
-)
-
-/**
- * Configuration entry for an asset. An asset is either a URL or a local file with its path and optionally its content
- * type
- */
-data class ResourceAssetConfig(
-    /**
-     * The URL to the asset
-     */
-    val url: String? = null,
-    /**
-     * The file for this asset
-     */
-    val file: String? = null,
-    /**
-     * The content type for the file
-     */
-    val contentType: String? = null
-)
-
-/**
- * A footer that should be displayed in the front-end
- */
-data class LinkFooterUrl(
-    /**
-     * The name (title) of the link, displayed in the front-end
-     */
-    val name: String,
-    /**
-     * The actual URL
-     */
-    val url: String
-)
-
-/**
- * Represents the contact information for a single person
- *
- * @since 0.2.0
- */
-data class LinkContactInformation(
-    /**
-     * The name of the person
-     */
-    val name: String,
-    /**
-     * Their email address
-     */
-    val email: String
-)
-
-/**
- * Tokens and secrets options
- *
- * @see LinkConfiguration.tokens
- */
-data class LinkTokens(
-    /**
-     * Discord bot token
-     */
-    val discordToken: String,
-    /**
-     * Discord OAuth Client ID
-     */
-    val discordOAuthClientId: String,
-    /**
-     * Discord OAuth Client Secret
-     */
-    val discordOAuthSecret: String,
-    /**
-     * Identity provider OpenID Connect / OAuth 2 Client Id
-     */
-    val idpOAuthClientId: String,
-    /**
-     * Identity provider OpenID Connect / OAuth 2 Client Secret
-     */
-    val idpOAuthSecret: String
-)
-
-/**
- * Configuration related to the Discord bot
- *
- * @see LinkConfiguration.discord
- */
-data class LinkDiscordConfig(
-    /**
-     * The URL the default welcome-I-don't-know-you message should redirect to.
-     */
-    val welcomeUrl: String?,
-    /**
-     * The prefix for commands. "e!" by default.
-     */
-    val commandsPrefix: String = "e!",
-    /**
-     * The list of EpiLink custom roles (list of [LinkDiscordRoleSpec] objects)
-     */
-    val roles: List<LinkDiscordRoleSpec> = listOf(),
-    /**
-     * The list of Discord servers that should be monitored by EpiLink
-     */
-    val servers: List<LinkDiscordServerSpec> = listOf(),
-    /**
-     * The default language to use in the bot
-     */
-    val defaultLanguage: String = "en",
-    /**
-     * The preferred languages list. Used for the first "change your language" prompt (only preferred languages are
-     * shown) + preferred languages are shown first in the e!lang list.
-     */
-    val preferredLanguages: List<String> = listOf(defaultLanguage)
-)
-
 
 /**
  * Check if a guild is monitored: that is, EpiLink knows how to handle it and is expected to do so.
  */
 fun LinkDiscordConfig.isMonitored(guildId: String): Boolean = servers.any { it.id == guildId }
 
-/**
- * A Discord EpiLink custom role
- *
- * @see LinkDiscordConfig.roles
- */
-data class LinkDiscordRoleSpec(
-    /**
-     * The name of this role
-     */
-    val name: String,
-    /**
-     * The human-friendly name of this role. Only used for display.
-     */
-    val displayName: String? = null,
-    /**
-     * The rule that should be used to determine this role.
-     */
-    val rule: String
-)
-
-/**
- * A Discord server that should be monitored by EpiLink
- *
- * @see LinkDiscordConfig.servers
- */
-data class LinkDiscordServerSpec(
-    /**
-     * The Discord ID of the server
-     */
-    val id: String,
-    /**
-     * True if a welcome message should be displayed, false otherwise
-     */
-    val enableWelcomeMessage: Boolean = true,
-    /**
-     * A Discord embed if a custom welcome message should be used, null (or absent) if the default message should be
-     * used.
-     */
-    val welcomeEmbed: DiscordEmbed? = null,
-    /**
-     * The role mapping. Keys are EpiLink roles, values are Discord roles.
-     */
-    val roles: Map<String, String>
-)
-
-/**
- * Privacy options for EpiLink, related to how some notifications should be handled.
- *
- * @see LinkConfiguration.privacy
- */
-data class LinkPrivacy(
-    /**
-     * True if automated accesses should send a notification
-     */
-    val notifyAutomatedAccess: Boolean = true,
-    /**
-     * True if human accesses should send a notification
-     */
-    val notifyHumanAccess: Boolean = true,
-    /**
-     * True if the name of the person who requested the identity access should be disclosed.
-     */
-    val discloseHumanRequesterIdentity: Boolean = false,
-    /**
-     * True if a ban should trigger a notification, false otherwise.
-     *
-     * Bans will not send a notification if banning someone who is not known.
-     */
-    val notifyBans: Boolean = true
-) {
-    /**
-     * Whether an access should be notified, based on this object's configuration
-     *
-     * @param automated True if the access was automated
-     */
-    fun shouldNotify(automated: Boolean): Boolean =
-        if (automated) notifyAutomatedAccess else notifyHumanAccess
-
-    /**
-     * Whether the identity of the requester should be disclosed, based on this object's configuration.
-     */
-    fun shouldDiscloseIdentity(automated: Boolean): Boolean =
-        if (automated) true else discloseHumanRequesterIdentity
-}
-
-/**
- * The legal configuration
- */
-data class LinkLegalConfiguration(
-    /**
-     * Terms of services, directly as a string
-     */
-    val tos: String? = null,
-    /**
-     * Terms of services as a path relative to the configuration file's location
-     */
-    val tosFile: String? = null,
-    /**
-     * Privacy policy, directly as a string
-     */
-    val policy: String? = null,
-    /**
-     * Privacy policy, as a path relative to the configuration file's location
-     */
-    val policyFile: String? = null,
-    /**
-     * The text shown below the "Remember who I am" checkbox in the front-end
-     */
-    val identityPromptText: String? = null
-)
-
-/**
- * Configuration for the identity provider
- */
-data class LinkIdProviderConfiguration(
-    /**
-     * Url of the issuer/authority (without the .well-known/openid-configuration
-     */
-    val url: String,
-    /**
-     * The name of the identity provider
-     */
-    val name: String,
-    /**
-     * Asset for the identity provider's icon, possibly null
-     */
-    val icon: ResourceAssetConfig?,
-    /**
-     * Enable backwards compat, for msft only.
-     */
-    val microsoftBackwardsCompatibility: Boolean = false
-)
 
 private val yamlKotlinMapper = ObjectMapper(YAMLFactory()).apply {
     registerModule(KotlinModule())
@@ -421,152 +120,3 @@ fun LinkConfiguration.isConfigurationSane(
     return report
 }
 
-/**
- * Check the coherence of the configuration with the available languages
- */
-fun LinkDiscordConfig.checkCoherenceWithLanguages(available: Set<String>): List<ConfigReportElement> {
-    val languages = available.joinToString(", ")
-    val report = mutableListOf<ConfigReportElement>()
-    if (defaultLanguage !in available) {
-        report += ConfigError(
-            true,
-            "The default language you chose ($defaultLanguage) is not available. The available languages are $languages"
-        )
-    }
-
-    val unavailablePreferredLanguages = preferredLanguages - available
-    if (unavailablePreferredLanguages.isNotEmpty()) {
-        report += ConfigError(
-            true,
-            "The preferred languages contain unavailable languages (${unavailablePreferredLanguages.joinToString(", ")}). The available languages are $languages"
-        )
-    }
-    if (defaultLanguage !in preferredLanguages) {
-        report += ConfigError(
-            true,
-            "The preferred languages list ${preferredLanguages.joinToString(", ")} must contain the default language ($defaultLanguage)."
-        )
-    }
-    return report
-}
-
-/**
- * Check the general Discord configuration
- */
-fun LinkDiscordConfig.check(): List<ConfigReportElement> {
-    val report = mutableListOf<ConfigReportElement>()
-    roles.map { it.name }.filter { it.startsWith("_") }.forEach {
-        report +=
-            ConfigWarning("A role was registered with the name ${it}, which starts with an underscore. Underscores are reserved for standard EpiLink roles like _known.")
-    }
-    return report
-}
-
-/**
- * Check the tokens' configuration
- */
-@Suppress("unused")
-fun LinkTokens.check(): List<ConfigReportElement> {
-    // Emit an error if using the default values from the sample config file
-    val report = mutableListOf<ConfigReportElement>()
-    if (discordOAuthClientId == "...") {
-        report += ConfigError(true, "discordOAuthClientId was left with its default value: please provide a client ID!")
-    }
-    if (discordOAuthSecret == "...") {
-        report += ConfigError(true, "discordOAuthSecret was left with its default value: please provide a secret!")
-    }
-    if (discordToken == "...") {
-        report += ConfigError(true, "discordToken was left with its default value: please provide a bot token!")
-    }
-    if (idpOAuthClientId == "...") {
-        report += ConfigError(true, "idpOAuthClientId was left with its default value: please provide a client ID!")
-    }
-    if (idpOAuthSecret == "...") {
-        report += ConfigError(true, "idpOAuthSecret was left with its default value: please provide a secret!")
-    }
-    return report
-}
-
-/**
- * Check the web server's configuration
- */
-fun LinkWebServerConfiguration.check(): List<ConfigReportElement> {
-    val reports = mutableListOf<ConfigReportElement>()
-    if (this.frontendUrl?.endsWith("/") == false) { // Equality check because left side can be null
-        reports += ConfigError(
-            true,
-            "The frontendUrl value in the server config must have a trailing slash (add a / at the end of your URL)"
-        )
-    }
-    return reports
-}
-
-/**
- * Check the coherence of the legal configuration
- */
-fun LinkLegalConfiguration.check(): List<ConfigReportElement> {
-    val rep = mutableListOf<ConfigReportElement>()
-    if (this.tos == null && this.tosFile == null) {
-        rep += ConfigWarning("No ToS provided. Please provide one using in the 'legal' part of the configuration, either as a file (tosFile) or directly as a string (tos)")
-    } else if (this.tos != null && this.tosFile != null) {
-        rep += ConfigError(
-            true,
-            "Your configuration file defines both a tos and a tosFile. Please only specify one of them, either only a file (tosFile) or a string (tos)"
-        )
-    }
-
-    if (this.policy == null && this.policyFile == null) {
-        rep += ConfigWarning("No privacy policy provided. Please provide one using the 'legal' part of the configuration, either as a file (policyFile) or directly as a string (policy)")
-    } else if (this.policy != null && this.policyFile != null) {
-        rep += ConfigError(
-            true,
-            "Your configuration file defines both a policy and a policyFile. Please only specify one of them, either only a file (policyFile) or a string (policy)"
-        )
-    }
-
-    if (this.identityPromptText == null) {
-        rep += ConfigWarning("No identity prompt text provided. Please provide one using the 'legal' part of the configuration (identityPromptText)")
-    }
-
-    return rep
-}
-
-/**
- * Report whether the rulebook and the config have some incoherent elements. Checks include:
- *
- * - Missing rules
- * - Missing role definitions
- * - Unused rules
- */
-private fun LinkDiscordConfig.checkCoherenceWithRulebook(rulebook: Rulebook): List<ConfigReportElement> {
-    val report = mutableListOf<ConfigReportElement>()
-    val roleNamesUsedInServers =
-        servers.map { it.roles.keys }.flatten().toSet() - StandardRoles.values().map { it.roleName }
-    val rolesDeclaredInRoles = roles
-    val rulesDeclared = rulebook.rules.keys
-
-    val roleNamesDeclaredInRoles = rolesDeclaredInRoles.map { it.name }.toSet()
-    val rolesMissingInRoles = roleNamesUsedInServers - roleNamesDeclaredInRoles
-    for (role in rolesMissingInRoles) {
-        report += ConfigError(
-            true,
-            "Role $role is referenced in a server config but is not defined in the Discord roles config"
-        )
-    }
-
-    val ruleNamesUsedInRoles = rolesDeclaredInRoles.map { it.rule }.toSet()
-    val missingRules = ruleNamesUsedInRoles - rulesDeclared
-    for (rule in missingRules) {
-        report += ConfigError(
-            true,
-            "Rule $rule is used in a Discord role config but is not defined in the rulebook"
-        )
-    }
-
-    val unusedRules = rulesDeclared - ruleNamesUsedInRoles
-    for (rule in unusedRules) {
-        report += ConfigWarning("Rule $rule is defined in the rulebook but is never used")
-    }
-
-    return report
-}
